FT2-37-3: Improved code readability and implemented DRY wherever possible. Description: - Condensed name validating functions into one to implement DRY. - Improved code indentation. - Specified @var, @param and @return for functions and classes. - Added proper comments for easy understanding.
