+++ script.js
@@ -2,116 +2,123 @@
 let errorFree = true;
 
 /**
- * Function to check for valid names and return
- * relevant Error message.
- * @params : (i)   string 'data' to be verified.
- *           (ii)  string 'regex' to validate 'data'.
- *           (iii) string 'field' conatins the field being verified.        
- * 
- * @returns : string
+ * Validates name fieldsn and find relevant errors.
+ *
+ * @params string data
+ *  Data to be verified.
+ * @param string regex
+ *  Regex to be validated against.
+ * @param string field
+ *  Field name which is being verified.
+ *
+ *  @return string
+ *  Stating relevant error message.
  */
 
 function checkInput(data, regex, field) {
-    let length = data.length;
-    if (length === 0) {
-        errorFree = false;
-        return `${field} is required.`;
-    }
-    else if (length < 2 || length > 25 ) {
-        errorFree = false;
-        return `${field} should be between 2 to 25 characters.`;
-    }
-    else if (!regex.test(data)) {
-        errorFree = false;
-        return "* Only alphabets, white-spaces and ' allowed.";
-    }
-    else {
-        return "*";
-    }
+  let length = data.length;
+  if (length === 0) {
+    // No data is entered.
+    errorFree = false;
+    return `${field} is required.`;
+  }
+  else if (length < 2 || length > 25 ) {
+    // Input data is not within specified length.
+    errorFree = false;
+    return `${field} should be between 2 to 25 characters.`;
+  }
+  else if (!regex.test(data)) {
+    // Input contains any forbidden characters.
+    errorFree = false;
+    return "* Only alphabets and white-spaces ";
+  }
+  return "*";
 }
 
 /**
-* Function to check if the input 'data' is a number.
-* @params : (i) 'data' to be verified.   
-*
-* @returns : boolean (true if 'data' is a number. otherwise false)
-*/
-
-function isNumber(data) { 
-    return /^-?[\d.]+(?:e-?\d+)?$/.test(data); 
-} 
+ * Function to check if the input 'data' is a number.
+ * @param string data
+ *  Data to be checked.
+ *
+ * @return bool
+ * true if data is a number, otherwise false.
+ */
+function isNumber(data) {
+    return /^-?[\d.]+(?:e-?\d+)?$/.test(data);
+}
 
 /**
-* Function to check if the input 'data' has valid 'Subject|Marks' pairs.
-* @params : (i)   string 'data' to be verified.   
-*           (ii)  string 'regex' to allow only specified characters.
-*
-* @returns : string containing the Error message(if any).
-*/
-
+ * Function to check if the input 'data' has valid 'Subject|Marks' pairs.
+ * @param string data
+ *  Data to be verified.
+ * @param string regex
+ *  Regex to be validated against.
+ *
+ * @return string
+ *  Containing the error message(if any).
+ */
 function checkSubjectMarks (data, regex){
-    if(data.length == 0) {
+  if(data.length == 0) {
+    errorFree = false;
+    return "* No subject|marks pairs entered";
+  }
+  else if (!regex.test(data)) {
+    errorFree = false;
+    return "* Only alphabets, digits and | allowed";
+  }
+  else{
+    let lines = data.split("\n");
+    let numberOfLines = lines.length;
+    for(i = 0; i < numberOfLines; i++){
+      subjectMarksPair = lines[i].split("|");
+      /* Trim any extra spaces in subject and marks. */
+      let subject = subjectMarksPair[0].trim();
+      let marks = subjectMarksPair[1].trim();
+      if (subjectMarksPair.length !== 2 || isNumber(subject) || !isNumber(marks)) {
+        /* More than one pair in a line, subject is numeric or marks is not numeric. */
         errorFree = false;
-        return "* No subject|marks pairs entered";
-    }
-    else if (!regex.test(data)) {
+        return "* Invalid Pairs";
+      }
+      else if(subjectMarksPair[1] > 100 ) {
         errorFree = false;
-        return "* Only alphabets, digits and | allowed";
+        return "* Marks cannot be greater than 100";
+      }
     }
-    else{
-        let lines = data.split("\n");
-        let numberOfLines = lines.length;
-        for(i = 0; i < numberOfLines; i++){
-            subjectMarksPair = lines[i].split("|");
-            if(subjectMarksPair.length !== 2 || isNumber(subjectMarksPair[0]) || !isNumber(subjectMarksPair[1])) {
-                errorFree = false;
-                return "* Invalid Pairs";
-            }
-            else if(subjectMarksPair[1] > 100 ) {
-                errorFree = false;
-                return "* Marks cannot be greater than 100";
-            }
-        }
-       
-    }
-    return "*";
+  }
+  return "*";
 }
 
 /**
  * Function to implement front-end validation
  * of the user inputs.
- * @params : No input parameters 
- * @returns : boolean (true if no errors).
+ *
+ * @return  bool
+ *  true if every field is properly filled, false otherwise.
  */
-
 function validateData() {
-    alert("Script is running");
-    /* Getting the value of the different user inputs */
-    let fname = document.getElementById('fname').value.trim();
-    let lname = document.getElementById('lname').value.trim();
-    let fullName = document.getElementById('fullName').value.trim();
-    errorFree = true;
+  /* Getting the value of the different user inputs */
+  let fname = document.getElementById('fname').value.trim();
+  let lname = document.getElementById('lname').value.trim();
+  errorFree = true;
 
-    /* Regex to check for valid names */
-    let nameRegex = /^[a-zA-Z\s']+$/;
-    let subjectMarksRegex = /^[a-zA-Z0-9|\s]+$/;
+  /* Regex to check for valid names */
+  let nameRegex = /^[a-zA-Z\s]+$/;
+  let subjectMarksRegex = /^[a-zA-Z0-9|\s]+$/;
 
-    /* Validation for First name */
-    let fnameError = document.getElementById('fnameError');
-    fnameError.innerHTML = checkInput(fname, nameRegex, "First name");
+  /* Validation for First name */
+  let fnameError = document.getElementById('fnameError');
+  fnameError.innerHTML = checkInput(fname, nameRegex, "First name");
 
-    /* Validation for Last name */
-    let lnameError = document.getElementById('lnameError');
-    lnameError.innerHTML = checkInput(lname, nameRegex, "Last name");
+  /* Validation for Last name */
+  let lnameError = document.getElementById('lnameError');
+  lnameError.innerHTML = checkInput(lname, nameRegex, "Last name");
 
-    /* Validation for 'Subject|Marks' pairs. */
-    let subjectMarks = document.getElementById("subjectMarks").value.trim();
-    let subjectMarksError = document.getElementById("subjectMarksError");
-    subjectMarksError.innerHTML = checkSubjectMarks(subjectMarks, subjectMarksRegex);
+  /* Validation for 'Subject|Marks' pairs. */
+  let subjectMarks = document.getElementById("subjectMarks").value.trim();
+  let subjectMarksError = document.getElementById("subjectMarksError");
+  subjectMarksError.innerHTML = checkSubjectMarks(subjectMarks, subjectMarksRegex);
 
-
-    /* Returning whether all the input fields are proper
-       and the form can be submitted or not. */
-    return errorFree;
+  // Returning whether all the input fields are properly filled
+  // and the form can be submitted or not.
+  return errorFree;
 }
-
