+++ style.css
@@ -1,139 +1,142 @@
-/* Resetting padding and margin to 0 and 
+/* Resetting padding and margin to 0 and
    setting the box-sizing property. */
 * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
 
 /* Setting height and background of body. */
 body {
-    background: rgb(34, 193, 195);
-    background: radial-gradient(circle, rgba(34, 193, 195, 1) 0%, rgba(253, 187, 45, 1) 100%);
-    min-height: 100svh;
+  background: rgb(34, 193, 195);
+  background: radial-gradient(circle, rgba(34, 193, 195, 1) 0%, rgba(253, 187, 45, 1) 100%);
+  min-height: 100svh;
 }
 
 /* Setting the width of container. */
 .container {
-    max-width: 800px;
-    margin: auto;
+  max-width: 800px;
+  margin: auto;
 }
 
 /* Styling the heading */
 .heading {
-    font-size: 50px;
-    margin: 30px 0px;
-    text-align: center;
+  font-size: 50px;
+  margin: 30px 0px;
+  text-align: center;
 }
 
 /* Styling labels for input fields. */
 label {
-    display: block;
-    font-size: 25px;
-    margin-bottom: 10px;
+  display: block;
+  font-size: 25px;
+  margin-bottom: 10px;
 }
 
 /* Styling error messages. */
 .error {
-    vertical-align: top;
-    color: red;
+  vertical-align: top;
+  color: red;
 }
 
 /* Styling the input fields */
 input {
-    outline: none;
-    border: none;
+  outline: none;
+  border: none;
 }
 
 /* Styling the text input fields. */
 input[type="text"] {
-    width: 95%;
-    margin: auto;
-    font-size: 25px;
-    padding: 5px 10px;
-    border-radius: 15px;
-    background-color: rgba(253, 253, 124, 0.808);
-    border: 2px solid black;
-    margin-bottom: 10px
+  width: 95%;
+  margin: auto;
+  font-size: 25px;
+  padding: 5px 10px;
+  border-radius: 15px;
+  background-color: rgba(253, 253, 124, 0.808);
+  border: 2px solid black;
+  margin-bottom: 10px
 }
 
 /* Styling the file input field. */
 input[type="file"] {
-    font-size: 20px;
-    padding: 5px 10px;
-    margin-bottom: 25px;
+  font-size: 20px;
+  padding: 5px 10px;
+  margin-bottom: 25px;
 }
 
 /* Styling the textarea for inputting Subject|Marks Pairs */
 textarea{
-    width: 95%;
-    margin: auto;
-    min-height:100px ;
-    background-color: rgba(253, 253, 124, 0.808);
-    border-radius: 20px;
-    border: 2px solid black;
-    resize: none;
-    outline: none;
-    margin-bottom: 25px;
-    padding: 8px 15px;
-    font-size: 20px;
-    overflow: hidden;
+  width: 95%;
+  margin: auto;
+  min-height:100px ;
+  background-color: rgba(253, 253, 124, 0.808);
+  border-radius: 20px;
+  border: 2px solid black;
+  resize: none;
+  outline: none;
+  margin-bottom: 25px;
+  padding: 8px 15px;
+  font-size: 20px;
+  overflow: hidden;
 }
 
 /* Stling the input button. */
 input[type="submit"] {
-    display: block;
-    font-size: 20px;
-    padding: 8px 20px;
-    border-radius: 10px;
-    margin-top: 25px;
+  display: block;
+  font-size: 20px;
+  padding: 8px 20px;
+  border-radius: 10px;
+  margin-top: 25px;
 }
 
 /* Hover effect on submit button. */
 input[type="submit"]:hover {
-    cursor: pointer;
-    background-color: rgb(230, 225, 58);
+  cursor: pointer;
+  background-color: rgb(230, 225, 58);
 }
 
 /* To show Full Name field is disabled. */
 #fullName {
-    cursor: not-allowed;
+  cursor: not-allowed;
 }
 
 /* Styling the displayed details. */
 .detail {
-    font-size: 25px;
-    text-transform: capitalize;
-    margin: 25px 0px;
-    text-align: center;
+  font-size: 25px;
+  text-transform: capitalize;
+  margin: 25px 0px;
+  text-align: center;
 }
 
 /* Styling the image displayed. */
 .showimage {
-    margin-top: 25px;
-    display: flex;
-    justify-content: center;
-    height: 250px;
+  margin-top: 25px;
+  display: flex;
+  justify-content: center;
+  height: 250px;
 }
 
 /* Stling the Subject|Marks table */
 table {
-    width: 100%;
-    border-collapse: collapse;
-    margin: 20px 0;
-    background-color: lightyellow; 
+  width: 100%;
+  border-collapse: collapse;
+  margin: 20px 0;
+  background-color: lightyellow;
 }
 
+/* Styling the Subject|Marks table header and data.*/
 th, td {
-    border: 1px solid #dddddd;
-    text-align: left;
-    padding: 8px;
+  border: 1px solid #dddddd;
+  text-align: left;
+  padding: 8px;
 }
 
+/* Giving a background color to the table headers.*/
 th {
-    background-color: rgba(253, 253, 124, 0.808);
+  background-color: rgba(253, 253, 124, 0.808);
 }
 
+/* Hover effect on table rows. */
 tr:hover {
-    background-color: rgba(253, 253, 124, 0.6);
+  background-color: rgba(253, 253, 124, 0.6);
 }
